removed the layout pages in favor of page
