+++ components/Anime/Streaming.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Streaming = () => {
+  return <div>Streaming</div>;
+};
+
+export default Streaming;
